face-pipeline/pipeline/embedder.py
import logging
from typing import List, Dict, Any, Optional, TYPE_CHECKING

    import numpy as np
    import PIL.Image

        import insightface
    import numpy as np

    # Placeholder: return zeros of shape (512,) float32

"""
Face Embedding Module

TODO: Implement face embedding/encoding using InsightFace
TODO: Support multiple embedding models (ArcFace, CosFace, etc.)
TODO: Add batch embedding for performance
TODO: Normalize embeddings for cosine similarity
TODO: Add embedding caching/memoization

POTENTIAL DUPLICATE: backend/app/services/face.py has embedding logic
"""

if TYPE_CHECKING:
logger = logging.getLogger(__name__)

# Global model singleton
_model = None


# ============================================================================
# Embedding Functions
# ============================================================================

def load_model() -> object:
    """
    Load and return the face embedding model (singleton pattern).

    Lazy-loads InsightFace model on first call and caches for subsequent calls.
    This avoids loading the model multiple times and saves memory.

    Returns:
        Loaded embedding model instance (e.g., InsightFace app)

    TODO: Initialize InsightFace app with buffalo_l model
    TODO: Set context ID from settings (CPU vs GPU)
    TODO: Configure detection size from settings
    TODO: Add error handling for model loading failures
    TODO: Support alternative embedding models
    TODO: Add model warmup (run on dummy input)

    Example implementation:
        app = insightface.app.FaceAnalysis(name='buffalo_l')
        app.prepare(ctx_id=-1, det_size=(640, 640))
        return app
    """
    global _model
    if _model is None:
        # TODO: Load model here
        # from config.settings import settings
        # import insightface
        # _model = insightface.app.FaceAnalysis(name=settings.detector_model)
        # _model.prepare(ctx_id=settings.detector_ctx_id, det_size=(settings.detector_size_width, settings.detector_size_height))
        pass
    return _model


def l2_normalize(embedding: "np.ndarray") -> "np.ndarray":
    """
    L2-normalize an embedding vector for cosine similarity.

    Normalizing embeddings to unit length allows using dot product
    instead of cosine similarity, which is faster for large-scale search.

    Args:
        embedding: Raw embedding vector (any length)

    Returns:
        L2-normalized embedding with unit length (norm = 1.0)

    TODO: Compute L2 norm (np.linalg.norm)
    TODO: Divide embedding by norm
    TODO: Handle zero vectors (return as-is or raise error)
    TODO: Validate input shape (1D array)

    Example:
        norm = np.linalg.norm(embedding)
        if norm > 0:
            return embedding / norm
        return embedding
    """
    # TODO: Implement L2 normalization
    pass


def embed(img_pil: "PIL.Image.Image") -> "np.ndarray":
    """
    Generate a 512-dimensional face embedding from a PIL image.

    Converts the face crop to an embedding vector suitable for similarity
    search and vector indexing. Uses InsightFace's ArcFace model.

    Args:
        img_pil: Face crop as PIL Image (RGB format)
                 Should be aligned and cropped face, not full image

    Returns:
        Normalized embedding vector as numpy array, shape (512,), dtype float32

    TODO: Load model using load_model()
    TODO: Convert PIL Image to numpy array (BGR for InsightFace)
    TODO: Run model inference to extract embedding
    TODO: Call l2_normalize() to normalize the embedding
    TODO: Ensure output is shape (512,) and dtype float32
    TODO: Add error handling for inference failures
    TODO: Add input validation (image size, format)

    Example flow:
        model = load_model()
        img_np = np.array(img_pil)[:, :, ::-1]  # RGB to BGR
        faces = model.get(img_np)
        if faces:
            embedding = faces[0].embedding
            return l2_normalize(embedding).astype(np.float32)
    """
    embedding = np.zeros(512, dtype=np.float32)

    # TODO: Call l2_normalize(embedding) once implemented
    # return l2_normalize(embedding)

    return embedding


class FaceEmbedder:
    """Face embedding/encoding service."""

    def __init__(
        self,
        model_name: str = "buffalo_l",
        ctx_id: int = -1,
        normalize: bool = True
    ):
        """
        Initialize face embedder.

        Args:
            model_name: InsightFace model name
            ctx_id: Context ID (-1 for CPU, 0+ for GPU)
            normalize: Whether to L2-normalize embeddings

        TODO: Load model lazily
        TODO: Support custom embedding models
        """
        self.model_name = model_name
        self.ctx_id = ctx_id
        self.normalize = normalize
        self._app = None

    def _load_model(self):
        """
        Load the face embedding model.

        TODO: Implement lazy loading
        TODO: Add error handling
        TODO: Support multiple model formats (ONNX, PyTorch, etc.)
        """
        pass

    def embed(self, face_crop: "np.ndarray") -> "np.ndarray":
        """
        Generate embedding for a face crop.

        Args:
            face_crop: Cropped face image as numpy array

        Returns:
            Embedding vector as numpy array

        TODO: Implement embedding extraction
        TODO: Add preprocessing (alignment, normalization)
        TODO: Handle edge cases (blur, occlusion)
        """
        pass

    async def embed_async(self, face_crop: "np.ndarray") -> "np.ndarray":
        """
        Async wrapper for face embedding.

        TODO: Run embedding in thread pool
        TODO: Handle cancellation
        """
        pass

    def embed_batch(self, face_crops: List["np.ndarray"]) -> List["np.ndarray"]:
        """
        Batch face embedding for multiple crops.

        TODO: Implement efficient batch processing
        TODO: Use GPU batching if available
        TODO: Add progress tracking
        """
        pass

    @staticmethod
    def compute_similarity(emb1: "np.ndarray", emb2: "np.ndarray") -> float:
        """
        Compute cosine similarity between two embeddings.

        TODO: Implement cosine similarity
        TODO: Support other distance metrics (Euclidean, etc.)
        """
        pass
