temp_requirements.txt
fastapi==0.115.0
gunicorn==22.0.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
psycopg[binary]==3.2.1
psycopg_pool==3.2.3
redis==5.0.7
celery==5.4.0
httpx==0.27.2
python-multipart==0.0.9
Pillow==10.4.0
imagehash==4.3.1

# Storage
boto3==1.34.162
minio==7.2.9

# Vector: local + prod
qdrant-client==1.10.1
pinecone-client==3.2.2

# Face stack (CPU default)
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
numpy==1.26.4

# Web crawling dependencies
aiofiles==23.2.1
aiohttp==3.10.11

# Optional: JavaScript rendering fallback
playwright==1.48.0

# Additional dependencies for NFR implementation
beautifulsoup4==4.12.2
python-dotenv==1.1.1

# System monitoring
psutil==6.1.0

# Content-addressed storage
blake3==1.0.7
